+++ src/classes/queue-base.ts
@@ -1,7 +1,12 @@
 import { EventEmitter } from 'events';
 import { QueueBaseOptions, RedisClient } from '../interfaces';
 import { MinimalQueue } from '../types';
-import { delay, DELAY_TIME_5, isNotConnectionError } from '../utils';
+import {
+  delay,
+  DELAY_TIME_5,
+  isNotConnectionError,
+  isRedisInstance,
+} from '../utils';
 import { RedisConnection } from './redis-connection';
 import { Job } from './job';
 import { KeysMap, QueueKeys } from './queue-keys';
@@ -60,7 +65,7 @@ export class QueueBase extends EventEmitter implements MinimalQueue {
 
     this.connection = new Connection(
       opts.connection,
-      opts.sharedConnection,
+      isRedisInstance(opts?.connection),
       opts.blockingConnection,
       opts.skipVersionCheck,
     );