+++ src/classes/worker.ts
@@ -200,7 +200,6 @@ export class Worker<
       name,
       {
         ...opts,
-        sharedConnection: isRedisInstance(opts.connection),
         blockingConnection: true,
       },
       Connection,